src/emailservice/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==5.2.0
    # via google-auth
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.1
    # via requests
google-api-core[grpc]==2.10.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-monitoring
    #   google-cloud-trace
    #   opencensus
google-api-python-client==2.58.0
    # via google-cloud-profiler
google-auth==2.11.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-profiler
google-auth-httplib2==0.1.0
    # via
    #   google-api-python-client
    #   google-cloud-profiler
google-cloud-core==2.3.2
    # via google-cloud-trace
google-cloud-monitoring==2.11.1
    # via opencensus-ext-stackdriver
google-cloud-profiler==4.0.0
    # via -r requirements.in
google-cloud-trace==0.24.2
    # via
    #   -r requirements.in
    #   opencensus-ext-stackdriver
googleapis-common-protos==1.56.4
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.48.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   grpcio-health-checking
    #   grpcio-status
    #   opencensus-ext-grpc
grpcio-health-checking==1.48.1
    # via -r requirements.in
grpcio-status==1.48.1
    # via google-api-core
httplib2==0.20.4
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.3
    # via requests
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.1
    # via jinja2
opencensus==0.11.0
    # via
    #   -r requirements.in
    #   opencensus-ext-grpc
    #   opencensus-ext-stackdriver
opencensus-context==0.1.3
    # via opencensus
opencensus-ext-grpc==0.7.2
    # via -r requirements.in
opencensus-ext-stackdriver==0.8.0
    # via -r requirements.in
proto-plus==1.22.1
    # via google-cloud-monitoring
protobuf==3.20.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-monitoring
    #   google-cloud-profiler
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpcio-health-checking
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.9
    # via httplib2
python-json-logger==2.0.4
    # via -r requirements.in
requests==2.28.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-profiler
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
opentelemetry-sdk==1.14.0
opentelemetry-api==1.14.0
opentelemetry-instrumentation-grpc
azure-monitor-opentelemetry-exporter
